Create README.md
From Nanodegree: "You will be provided with a design mockup as a PDF-file and must replicate that design in HTML and CSS. You will develop a responsive website that will display images, descriptions and links to each of the portfolio projects you will complete throughout the course of the Front-End Web Developer Nanodegree."  I have used the following in the process of creating the mockup site:  Google, Stack Overflow, w3 Schools, RollTide.com, git and github.  I tried to create a site that would exceed the mockup provided in the class.